+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Not released]
+
+Behaviour changes:
+ * The old function `SetSignature` in object and container tokens was renamed into `SignIssued` (#718)
+ * New function `SetSignature` works as setter only in object and container tokens (#718)
+
 ## [1.0.0-rc.13] - 2025-03-06
 
 The key change is the removal of `github.com/nspcc-dev/neofs-api-go/v2` module (#667).